backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.34
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
PyJWT==2.9.0
pandas==2.2.2
scikit-learn==1.5.1
